+++ src/proxy/http/mod.rs
@@ -1,9 +1,10 @@
 pub mod client;
 pub(super) mod glue;
 pub mod h1;
+pub mod normalize_uri;
+pub mod orig_proto;
 pub mod router;
 pub mod upgrade;
-pub mod orig_proto;
 
 pub use self::client::{Client, Error as ClientError};
 pub use self::glue::HttpBody as Body;
